+++ test/example/c-api-kitchen-sink.c
@@ -2019,14 +2019,18 @@ void test_for_each() {
                       BinaryenTypeInt32(),
                       0,
                       makeInt32(module, expected_offsets[1]));
-
+    assert(BinaryenGetDataSegment(module, segmentNames[0]) != NULL);
+    assert(BinaryenGetDataSegment(module, "NonExistentSegment") == NULL);
     for (i = 0; i < BinaryenGetNumMemorySegments(module); i++) {
       char out[15] = {};
-      assert(BinaryenGetMemorySegmentByteOffset(module, segmentNames[i]) ==
+      BinaryenDataSegmentRef segment = BinaryenGetDataSegmentByIndex(module, i);
+      assert(segment != NULL);
+      assert(BinaryenDataSegmentGetName(segment) != NULL);
+      assert(BinaryenGetMemorySegmentByteOffset(module, segment) ==
              expected_offsets[i]);
-      assert(BinaryenGetMemorySegmentByteLength(module, segmentNames[i]) ==
-             segmentSizes[i]);
-      BinaryenCopyMemorySegmentData(module, segmentNames[i], out);
+      assert(BinaryenGetMemorySegmentByteLength(segment) == segmentSizes[i]);
+      assert(BinaryenGetMemorySegmentPassive(segment) == segmentPassives[i]);
+      BinaryenCopyMemorySegmentData(segment, out);
       assert(0 == strcmp(segmentDatas[i], out));
     }
   }